\name{naturalistDB-package}
\alias{naturalistDB-package}
\alias{naturalistDB}
\docType{package}
\title{
\packageTitle{naturalistDB}
}
\description{
\packageDescription{naturalistDB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{naturalistDB}
\packageIndices{naturalistDB}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{naturalistDB}

Maintainer: \packageMaintainer{naturalistDB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
